refactor: modify routes & add admin authentication
