+++ controllers/admin-controller.js
@@ -1,0 +1,7 @@
+const adminServices = require('../services/admin-services')
+const adminController = {
+  getUsers: (req, res, next) => {
+    adminServices.getUsers(req, (err, data) => err ? next(err) : res.json(data))
+  }
+}
+module.exports = adminController
