+++ controllers/tweet-controller.js
@@ -1,0 +1,5 @@
+const tweetServices = require('../services/tweet-services')
+const tweetController = {
+  // Controller action here
+}
+module.exports = tweetController
