+++ middleware/api-auth.js
@@ -7,8 +7,7 @@ const authenticated = (req, res, next) => {
   })(req, res, next)
 }
 const authenticatedAdmin = (req, res, next) => {
-  if (req.user && (req.user.role ===
-    true)) return next()
+  if (req.user && (req.user.role === 'admin')) return next()
   return res.status(403).json({ status: 'error', message: 'permission denied' })
 }
 module.exports = {
