+++ routes/modules/followship.js
@@ -1,9 +1,9 @@
 const express = require('express')
 const router = express.Router()
-const userController = require('../../controllers/user-controller')
+const followshipController = require('../../controllers/followship-controller')
 
-router.post('/', userController.addFollowing)
-router.delete('/:userId', userController.removeFollowing)
-router.get('/', userController.getTopUsers)
+router.post('/', followshipController.addFollowing)
+router.delete('/:followingId', followshipController.removeFollowing)
+router.get('/', followshipController.getTopUsers)
 
 module.exports = router
