+++ routes/modules/tweet.js
@@ -1,0 +1,7 @@
+const express = require('express')
+const router = express.Router()
+const tweetController = require('../../controllers/tweet-controller')
+
+// add routes here
+
+module.exports = router
