+++ services/followship-services.js
@@ -1,0 +1,55 @@
+const { User, Followship } = require('./../models')
+const sequelize = require('sequelize')
+const followshipServices = {
+  getTopUsers: (req, cb) => {
+    const limit = Number(req.query.top)
+    return Followship.findAll({
+      include: [{ model: User, as: 'followingUser' }],
+      attributes: ['followingId', [sequelize.fn('COUNT', 'followingId'), 'count']],
+      order: [[sequelize.literal('count'), 'DESC']],
+      group: ['followingId'],
+      limit
+    })
+      .then(popularUsers => cb(null, popularUsers))
+      .catch(err => cb(err))
+  },
+  addFollowing: (req, cb) => {
+    const { id } = req.body
+    Promise.all([
+      User.findByPk(id),
+      Followship.findOne({
+        where: {
+          followerId: req.user.id,
+          followingId: req.body.id
+        }
+      })
+    ])
+      .then(([user, followship]) => {
+        if (!user) throw new Error("User didn't exist!")
+        if (followship) throw new Error('You are already following this user!')
+        return Followship.create({
+          followerId: req.user.id,
+          followingId: req.body.id
+        })
+      })
+      .then(followship => {
+        cb(null, { status: 'success', followship })
+      })
+      .catch(err => cb(err))
+  },
+  removeFollowing: (req, cb) => {
+    Followship.findOne({
+      where: {
+        followerId: req.user.id,
+        followingId: req.params.followingId
+      }
+    })
+      .then(followship => {
+        if (!followship) throw new Error("You haven't followed this user!")
+        return followship.destroy()
+      })
+      .then(deletedFollowship => cb(null, { status: 'success', deletedFollowship }))
+      .catch(err => cb(err))
+  }
+}
+module.exports = followshipServices
