+++ services/tweet-services.js
@@ -1,0 +1,5 @@
+const { User, Tweet, Reply, Like } = require('./../models')
+const tweetServices = {
+  // add services here
+}
+module.exports = tweetServices
